Django/requirements.txt
absl-py==1.4.0
asgiref==3.5.2
astunparse==1.6.3
av==10.0.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cmake==3.25.2
contourpy==1.0.7
cycler==0.11.0
dataclasses==0.6
django==4.1
dlib-19.22.99-cp39-cp39-win_amd64.whl
face-alignment==1.3.5
flatbuffers==1.12
fonttools==4.38.0
future==0.18.2
gast==0.4.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==3.8.0
idna==2.10
imageio==2.25.0
imageio-ffmpeg==0.4.8
importlib-metadata==2.0.0
keras==2.9.0
keras-preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
llvmlite==0.39.1
markdown==3.3.3
matplotlib==3.6.3
networkx==3.0
numba==0.56.4
numpy==1.23.5
oauthlib==3.1.0
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
opt-einsum==3.3.0
pillow==9.4.0
pip==22.3.1
protobuf==3.13.0
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pywavelets==1.4.1
pyyaml==4.2b4
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.19.3
scipy==1.10.0
setuptools==65.6.3
shapely==1.7.1
six==1.15.0
sqlparse==0.4.3
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tifffile==2023.2.3
tqdm==4.64.1
typing-extensions==3.7.4.3
urllib3==1.25.11
werkzeug==1.0.1
wheel==0.37.1
wincertstore==0.2
wrapt==1.14.1
zipp==3.4.0
-f https://download.pytorch.org/whl/cu110/torch_stable.html torch==1.7.1+cu110 torchaudio==0.7.2
torchvision==0.8.2
fastapi
uvicorn